pyproject.toml
[tool.poetry]
name = "poetry config virtualenvs-in-project true"
version = "0.1.0"
description = ""
authors = ["V_A_N <V_A_N-1_94@mail.ru>"]
readme = "README.md"
packages = [{include = "poetry config virtualenvs"}]

[tool.poetry.dependencies]
python = "^3.11"
xlrd = "^2.0.1"
pandas = "^2.1.3"
requests = "^2.31.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.lint.dependencies]
flake8 = "^6.1.0"
mypy = "^1.7.1"
isort = "^5.12.0"
black = "^23.11.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'